refactor(frontend): move workspace domain types into workspaces module
