+++ frontend/src/interfaces/index.ts
@@ -1,30 +1,5 @@
 export * from "@/modules/documents/domain";
-
-interface WorkspaceMeta {
-   workspace_uuid: string;
-   type: "workspace";
-   creatorNodeId: string;
-   creatorUserId: string;
-   creatorName?: string;
-   name: string;
-   password_hash?: string;
-   is_public: boolean;
-   created_at: string;
-}
-
-export interface WorkspaceRequest {
-   uuid: string;
-   meta: WorkspaceMeta;
-}
-
-export interface WorkspaceCreateResponse {
-   uuid: string;
-   cid: string;
-}
-
-export interface Workspace extends WorkspaceRequest {
-   cid: string;
-}
+export * from "@/modules/workspaces/domain";
 
 interface ChatMessageMeta {
    type: "chat";
