+++ frontend/src/modules/workspaces/domain/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./workspace.types";
