+++ frontend/src/modules/workspaces/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./domain";
